backend/requirements.txt
alembic==1.15.2
email-validator==2.2.0
fastapi==0.115.12
httpx==0.28.1
passlib[bcrypt]==1.7.4
py2neo==2021.2.4
PyMySQL==1.1.1
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-jose[cryptography]==3.4.0
pytest==8.3.5
requests==2.32.3
sqlalchemy==2.0.40
uvicorn[standard]==0.34.1
